requirements.txt
ultralytics>=8.3.0
roboflow>=1.1.0
opencv-python
pyyaml
